% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP-prm.R
\name{PRMap}
\alias{PRMap}
\title{Analytical power calculation for PRM design problem}
\usage{
PRMap(alpha = 0.05, n = 20, m = 20, s2w = 4.44, s2b = 1.82,
  mu = 1.01, rho = 3, REML = TRUE)
}
\arguments{
\item{alpha}{desired significance level. Default to 0.05}

\item{n}{number of subjects}

\item{m}{number of repeated measures for each subject}

\item{s2w}{within subject variation}

\item{s2b}{between subject variation}

\item{mu}{mean measure difference}

\item{rho}{null threshold of acceptable RMS value}

\item{REML}{using REML instead of MLE. Default to TRUE.}
}
\value{
\describe{
  \item{xq}{ computed quantile for the RMS test statistic }
  \item{size}{ computed actual type I error }
  \item{pwr}{ computed power }
  \item{par0}{ estimated null parameter values }
}
}
\description{
Due to its nature of composite null hypothesis, it is not trivial to derive the analytical power calculation. We develop an average likelihood
based approach.
}
\references{
Bai,Y., Wang,Z., Lystig,T.C., and Wu,B. (2018) Statistical test with sample size and power calculation for paired repeated measures designs of method comparison studies.
}
