% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP-prm.R
\name{PRMas}
\alias{PRMas}
\title{Analytical sample size calculation for PRM design problem}
\usage{
PRMas(pwr = 0.8, alpha = 0.05, m, s2w, s2b, mu, rho, REML = TRUE)
}
\arguments{
\item{pwr}{desired power. Default to 0.8}

\item{alpha}{desired significance level. Default to 0.05}

\item{m}{number of repeated measures for each subject}

\item{s2w}{within subject variation}

\item{s2b}{between subject variation}

\item{mu}{mean measure difference}

\item{rho}{null threshold of acceptable QMS value}

\item{REML}{using REML instead of MLE. Default to TRUE.}
}
\value{
\describe{
  \item{n}{ computed number of units }
  \item{pwr}{ actual power under the computed sample size }
}
}
\description{
Due to discrete nature of sample size, the power is not a continuous function. This function utilizes the analytical power calculation function PRSap()
to compute an approximate sample size n under a given m (number of measures per unit), which has power closest to the desired power.
}
\references{
Bai,Y., Wang,Z., Lystig,T.C., and Wu,B. (2018) Statistical test with sample size and power calculation for paired repeated measures designs of method comparison studies.
}
