% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP-prm.R
\name{PRMsp}
\alias{PRMsp}
\title{Sample size and power calculation for PRM design}
\usage{
PRMsp(alpha = 0.05, n = 20, m = 20, s2w = 4.44, s2b = 1.82, rho = 3,
  rhoa = 2.7)
}
\arguments{
\item{alpha}{desired significance level. Default to 0.05}

\item{n}{number of subjects}

\item{m}{number of repeated measures for each subject}

\item{s2w}{within subject variation}

\item{s2b}{between subject variation}

\item{rho}{null threshold of acceptable RMS value}

\item{rhoa}{actual RMS value}
}
\value{
\describe{
  \item{xq}{ computed quantile for the RMS test statistic }
  \item{size}{ computed actual type I error }
  \item{pwr}{ computed power }
}
}
\description{
Sample size and power calculation for PRM design
}
\references{
Bai,Y., Wu,B. and ... (2018) Statistical test with sample size and power calculation for paired repeated measures designs of method comparison studies.
}
