% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisum-prob.R
\name{pchisum}
\alias{pchisum}
\title{Compute the tail probability of weighted sum of non-central chi-square random variables}
\usage{
pchisum(Q.all, lambda, h = rep(1, length(lambda)), delta = rep(0,
  length(lambda)), acc = 0.01, lim = 1e+07)
}
\arguments{
\item{Q.all}{test statistics}

\item{lambda}{mixing coefficients}

\item{h}{DFs of chi-square dist}

\item{delta}{non-centrality parameters of chi-square dist}

\item{acc}{relative error bound}

\item{lim}{maximum number of integration terms}
}
\value{
tail probability
}
\description{
Use Davies' method and apply a relative error bound (avoid redundant computations): efficient and accurate.
}
\references{
Wu,B., Guan,W., Pankow,J.S. (2016) On efficient and accurate calculation of significance p-values for sequence kernel association test of variant set. \emph{Annals of human genetics}, 80(2), 123-135.

Bai,Y., Wang,Z., Lystig,T.C., and Wu,B. (2018) Statistical test with sample size and power calculation for paired repeated measures designs of method comparison studies.
}
