% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SP-prm.R
\name{PRMmcp}
\alias{PRMmcp}
\title{Monte Carlo power calculation for PRM design problem}
\usage{
PRMmcp(alpha = 0.05, n, m, s2w, s2b, mu, rho, REML = TRUE, B = 1000)
}
\arguments{
\item{alpha}{desired significance level. Default to 0.05}

\item{n}{number of subjects}

\item{m}{number of repeated measures for each subject}

\item{s2w}{within subject variation}

\item{s2b}{between subject variation}

\item{mu}{mean measure difference}

\item{rho}{null threshold of acceptable QMS value}

\item{REML}{using REML instead of MLE. Default to TRUE.}

\item{B}{total number of Monte Carlo simulations. Default to 1000.}
}
\value{
\describe{
  \item{pwr}{ computed power }
  \item{p.value}{ p-values from Monte Carlo simulations }
}
}
\description{
Due to its nature of composite null hypothesis, it is not trivial to derive the analytical power calculation. 
Monte Carlo simulation offers a straightforward approach by generating p-value random numbers.
}
\references{
Bai,Y., Wang,Z., Lystig,T.C., and Wu,B. (2018) Statistical test with sample size and power calculation for paired repeated measures designs of method comparison studies.
}
